Move Map to the fragment
