+++ mobile/src/main/java/cz/uhk/fim/kikm/wearnavigation/model/database/helpers/DatabaseDataLoader.java
@@ -1,51 +1,46 @@
 package cz.uhk.fim.kikm.wearnavigation.model.database.helpers;
 
-import android.content.Context;
 import android.os.AsyncTask;
 
-public class DatabaseDataLoader<T> extends AsyncTask<Integer, Void, T> {
+import java.util.List;
+
+import cz.uhk.fim.kikm.wearnavigation.model.database.Fingerprint;
+
+public class DatabaseDataLoader extends AsyncTask<Integer, Void, List<Fingerprint>> {
 
     // Modes of this loader
     public final static int MODE_FINGERPRINT = 0;
 
-    // Database crud to get access to the data in the SQL
-    private DatabaseCRUD mDatabase;
-    // Interface to communicate with the context
-    private DatabaseDataInterface mInterface;
-    // Mode of the running download
-    private Integer mMode = -1;
+    private DatabaseCRUD mDatabase;             // Database crud to get access to the data in the SQL
+    private DatabaseDataInterface mInterface;   // Interface to communicate with the context
+    private Integer mMode = -1;                 // Mode of the running download
 
     /**
      * Constructor that checks for the interface and sets proper variables.
      *
-     * @param context from which is this task loaded.
+     * @param database instance so we do not have to use Context
+     * @param pInterface to communicate via
      */
-    public DatabaseDataLoader(Context context) {
-        // Checks and loads the interface
-        try {
-            mInterface = (DatabaseDataInterface) context;
-        } catch (Exception e) {
-            throw new ClassCastException(context.getClass()
-                    + " must implement IndividualResultListener");
-        }
-
-        // Creates database CRUD helper instance
-        mDatabase = new DatabaseCRUD(context);
+    public DatabaseDataLoader(DatabaseCRUD database, DatabaseDataInterface pInterface) {
+        mDatabase = database;
+        mInterface = pInterface;
     }
 
     @Override
-    protected T doInBackground(Integer... params) {
-        mMode = params[0];
-        switch (mMode) {
-            case MODE_FINGERPRINT:
-                // Return all fingerprints from the database
-                return (T) mDatabase.getAllFingerprints(false);
+    protected List<Fingerprint> doInBackground(Integer... params) {
+        if(mDatabase != null && mInterface != null) {
+            mMode = params[0];
+            switch (mMode) {
+                case MODE_FINGERPRINT:
+                    // Return all fingerprints from the database
+                    return mDatabase.getAllFingerprints(false);
+            }
         }
         return null;
     }
 
     @Override
-    protected void onPostExecute(T result) {
+    protected void onPostExecute(List<Fingerprint> result) {
         if(result != null) {
             switch (mMode) {
                 case MODE_FINGERPRINT:
