+++ mobile/src/main/res-scan/values/strings.xml
@@ -8,6 +8,8 @@
     <!-- Actions for New Markers -->
     <string name="amn_title">New Fingerprint</string>
     <string name="amn_position">At position (%1$d:%2$d)</string>
-    
-    <string name="am_scan_already_running">There is a scan running at this moment.</string>
+
+    <!-- Fragment Map -->
+    <string name="fma_scan_already_running">There is a scan running at this moment.</string>
+    <string name="fma_scan_load_error">Could not load job data.\nNot able to create scan.</string>
 </resources>
